apps/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
motor==3.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyMuPDF==1.25.1
pillow==11.0.0
pytesseract==0.3.13
filetype==1.2.0
